Добаления администрирования. Рефакторинг директорий
