+++ source/webapp/lib/db.py
@@ -4,9 +4,9 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker, scoped_session
 
-import lib.config
+import webapp.lib.config
 
-engine = create_engine(lib.config.URL_API_KEY)
+engine = create_engine(webapp.lib.config.URL_API_KEY)
 db_session = scoped_session(sessionmaker(bind=engine))
 
 Base = declarative_base()
