+++ source/webapp/templates/admin/index.html
@@ -1,0 +1,1 @@
+{% extends "base.html" %} {% block content %} Администрирование {% endblock %}
