+++ source/webapp/templates/intro/index.html
@@ -1,0 +1,7 @@
+{% extends "base.html" %} {% block content %} {% with messages =
+get_flashed_messages() %} {% if messages %}
+<div class="alert alert-warning" role="alert">
+  {% for message in messages %} {{ message }}<br />
+  {% endfor %}
+</div>
+{% endif %} {% endwith %} {% endblock %}
