--- source/webapp/templates/login.html
@@ -1,36 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>{{ page_title }}</title>
-    <!--Bootstrap link css-->
-    <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
-    />
-    <link
-      rel="stylesheet"
-      href="{{ url_for('static', filename='css/styles.css') }}"
-    />
-  </head>
-  <body>
-    <div class="conteiner">
-      <h2 class="text-center">{{ page_title }}</h2>
-      {% with messages = get_flashed_messages() %} {% if messages %}
-      <div class="alert alert-warning" role="alert">
-        {% for message in messages %} {{ message }}<br />
-        {% endfor %}
-      </div>
-      {% endif %} {% endwith %}
-      <form action="{{ url_for('process_login') }}" method="POST">
-        {{ form.hidden_tag() }}
-        <p>{{ form.email.label }}<br />{{ form.email() }}</p>
-        <p>{{ form.password.label }}<br />{{ form.password() }}</p>
-        <p>{{ form.submit }}</p>
-      </form>
-    </div>
-    <!--Bootstrap link js-->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
-  </body>
-</html>
