+++ source/webapp/templates/user/login.html
@@ -1,0 +1,18 @@
+{% extends "base.html" %} {% block content %} {% with messages =
+get_flashed_messages() %} {% if messages %}
+<div class="alert alert-warning" role="alert">
+  {% for message in messages %} {{ message }}<br />
+  {% endfor %}
+</div>
+{% endif %} {% endwith %}
+<form action="{{ url_for('user.process_login') }}" method="POST">
+  {{ form.hidden_tag() }}
+  <div class="form-group">{{ form.email.label }} {{ form.email() }}</div>
+  <div class="form-group">{{ form.password.label }} {{ form.password() }}</div>
+  <div class="form-group form-check">
+    {{ form.remember_me() }} {{
+    form.remember_me.label(class_="form-check-label") }}
+  </div>
+  {{ form.submit }}
+</form>
+{% endblock %}
